pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glm-experiments"
version = "0.0.1"
description = "gLM experiments"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Gonzalo Benegas", email = "gonzalo.benegas@openathena.ai"},
]
keywords = ["pytorch", "lightning", "hydra", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "torchmetrics>=0.11.4",
    "hydra-core>=1.3.0,<2.0.0",
    "hydra-colorlog>=1.2.0",
    "hydra-optuna-sweeper>=1.2.0",
    "rootutils",
    "rich>=13.0.0",
    "datasets>=2.14.0",
    "transformers>=4.30.0",
    "biopython>=1.81",
    "zstandard>=0.18.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
train_command = "glm_experiments.train:main"
eval_command = "glm_experiments.eval:main"

[tool.hatchling.build]
packages = ["glm_experiments"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
